+++ changelog.md
@@ -24,7 +24,9 @@ All notable changes to this project will be documented in this file.
     - `NotEmptyCollection` and `NotEmptySet`
       ([#7](https://github.com/kotools/types/issues/7))
     - `NotEmptyMutableList`
-      ([#12](https://github.com/kotools/types/issues/12)).
+      ([#12](https://github.com/kotools/types/issues/12))
+    - `NotEmptyMutableSet`
+      ([#13](https://github.com/kotools/types/issues/13)).
 - Overload comparisons of the following types in `kotools.types.number`:
     - `NonZeroInt` ([#33](https://github.com/kotools/types/issues/33))
     - `PositiveInt` ([#34](https://github.com/kotools/types/issues/34))